candle_core::metal_backend

Struct MetalDevice

pub struct MetalDevice { /* private fields */ }

Implementations

impl MetalDevice

pub fn compile( &self, func_name: &'static str, kernel: Kernel, ) -> Result<ComputePipelineState>

pub fn id(&self) -> DeviceId

pub fn metal_device(&self) -> &Device

pub fn command_buffer(&self) -> Result<CommandBuffer>

pub fn wait_until_completed(&self) -> Result<()>

pub fn kernels(&self) -> &Kernels

pub fn device(&self) -> &Device

pub fn new_buffer( &self, element_count: usize, dtype: DType, name: &str, ) -> Result<Arc<Buffer>>

Creates a new buffer (not necessarily zeroed). The buffer is MTLPrivate This means the buffer data cannot be read on the CPU directly.

[name] is only used to keep track of the resource origin in case of bugs

pub fn new_buffer_managed(&self, size: NSUInteger) -> Result<Arc<Buffer>>

Creates a new buffer (not necessarily zeroed). The buffer is MTLManaged This means the buffer can be read on the CPU but will require manual synchronization when the CPU memory is modified Used as a bridge to gather data back from the GPU

pub fn new_buffer_with_data<T>(&self, data: &[T]) -> Result<Arc<Buffer>>

Creates a new buffer from data. The buffer is MTLManaged

Does not require synchronization, as newBufferWithBytes allocates the buffer and copies over the existing data before returning the MTLBuffer.

pub fn allocate_zeros(&self, size_in_bytes: usize) -> Result<Arc<Buffer>>

pub fn capture<P: AsRef<Path>>(&self, path: P) -> Result<()>

Create a metal GPU capture trace on [path].