candle_core::quantized::k_quants

Trait GgmlType

pub trait GgmlType:
    Sized
    + Clone
    + Send
    + Sync {
    type VecDotType: GgmlType;

    const DTYPE: GgmlDType;
    const BLCK_SIZE: usize;

    // Required methods
    fn to_float(xs: &[Self], ys: &mut [f32]) -> Result<()>;
    fn from_float(xs: &[f32], ys: &mut [Self]) -> Result<()>;
    fn vec_dot(n: usize, xs: &[Self], ys: &[Self::VecDotType]) -> Result<f32>;
    fn vec_dot_unopt(
        n: usize,
        xs: &[Self],
        ys: &[Self::VecDotType],
    ) -> Result<f32>;

    // Provided method
    fn zeros() -> Self { ... }
}

Required Associated Constants

const DTYPE: GgmlDType

const BLCK_SIZE: usize

Required Associated Types

type VecDotType: GgmlType

Required Methods

fn to_float(xs: &[Self], ys: &mut [f32]) -> Result<()>

fn from_float(xs: &[f32], ys: &mut [Self]) -> Result<()>

fn vec_dot(n: usize, xs: &[Self], ys: &[Self::VecDotType]) -> Result<f32>

Dot product used as a building block for quantized mat-mul. n is the number of elements to be considered.

fn vec_dot_unopt(n: usize, xs: &[Self], ys: &[Self::VecDotType]) -> Result<f32>

Generic implementation of the dot product without simd optimizations.

Provided Methods

fn zeros() -> Self

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.