candle_core

Struct Var

pub struct Var(/* private fields */);

A variable is a wrapper around a tensor, however variables can have their content modified whereas tensors are immutable.

Implementations

impl Var

pub fn zeros<S: Into<Shape>>( shape: S, dtype: DType, device: &Device, ) -> Result<Self>

pub fn ones<S: Into<Shape>>( shape: S, dtype: DType, device: &Device, ) -> Result<Self>

pub fn from_tensor(t: &Tensor) -> Result<Self>

pub fn rand_f64<S: Into<Shape>>( lo: f64, up: f64, s: S, dtype: DType, device: &Device, ) -> Result<Self>

pub fn randn_f64<S: Into<Shape>>( mean: f64, std: f64, s: S, dtype: DType, device: &Device, ) -> Result<Self>

pub fn rand<S: Into<Shape>, T: FloatDType>( lo: T, up: T, s: S, device: &Device, ) -> Result<Self>

pub fn randn<S: Into<Shape>, T: FloatDType>( mean: T, std: T, s: S, device: &Device, ) -> Result<Self>

pub fn new<A: NdArray>(array: A, device: &Device) -> Result<Self>

Creates a new tensor on the specified device using the content and shape of the input. This is similar to new but the resulting tensor is a variable.

pub fn from_vec<S: Into<Shape>, D: WithDType>( data: Vec<D>, shape: S, device: &Device, ) -> Result<Self>

pub fn from_slice<S: Into<Shape>, D: WithDType>( array: &[D], shape: S, device: &Device, ) -> Result<Self>

pub fn as_detached_tensor(&self) -> Tensor

pub fn as_tensor(&self) -> &Tensor

pub fn into_inner(self) -> Tensor

Consumes this Var and return the underlying tensor.

pub fn set(&self, src: &Tensor) -> Result<()>

Sets the content of the inner tensor, this does not require a mutable reference as inner mutability is used.